requirements.txt
flask
instaloader
pillow
pytesseract
nltk
